% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Functions.R
\name{Allocator}
\alias{Allocator}
\title{Allocator}
\usage{
Allocator(tau.ind.hat, beta.hat, tau.ind.true = tau.ind.true, p = p)
}
\arguments{
\item{tau.ind.hat}{indices of estimated PoIs}

\item{beta.hat}{empirical slope coefficients}

\item{tau.ind.true}{indices of true PoIs}

\item{p}{number of discretization points}
}
\description{
Allocates the estimated PoIs to the true PoIs
}
