% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Functions.R
\name{BM.SIM}
\alias{BM.SIM}
\title{BM.SIM}
\usage{
BM.SIM(a, b, p)
}
\arguments{
\item{a}{lower interval boundary}

\item{b}{upper interval boundary}

\item{p}{length of the path}
}
\description{
This function allows to simulate Brownian motion processes.
}
