% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Functions.R
\name{FUN_LMcK}
\alias{FUN_LMcK}
\title{FUN_LMcK}
\usage{
FUN_LMcK(Y = Y, X.mat = X.mat)
}
\arguments{
\item{Y}{dependent variable (n-vector)}

\item{X.mat}{pxn matrix of discretized functional predictors}
}
\description{
This function is an implementation of the estimation procedure described in 
'Logistic Regression With Brownian-Like Predictors' 
by Martin Lindquist and Ian McKeague
}
